+++ BulletSharp/Collision/CapsuleShape.cs
@@ -10,21 +10,21 @@ namespace BulletSharp
 		{
 		}
 
-		public CapsuleShape(float radius, float height)
+		public CapsuleShape(double radius, double height)
 			: base(btCapsuleShape_new(radius, height))
 		{
 		}
 
-		public float HalfHeight => btCapsuleShape_getHalfHeight(Native);
+		public double HalfHeight => btCapsuleShape_getHalfHeight(Native);
 
-		public float Radius => btCapsuleShape_getRadius(Native);
+		public double Radius => btCapsuleShape_getRadius(Native);
 
 		public int UpAxis => btCapsuleShape_getUpAxis(Native);
 	}
 
 	public class CapsuleShapeX : CapsuleShape
 	{
-		public CapsuleShapeX(float radius, float height)
+		public CapsuleShapeX(double radius, double height)
 			: base(btCapsuleShapeX_new(radius, height))
 		{
 		}
@@ -32,7 +32,7 @@ namespace BulletSharp
 
 	public class CapsuleShapeZ : CapsuleShape
 	{
-		public CapsuleShapeZ(float radius, float height)
+		public CapsuleShapeZ(double radius, double height)
 			: base(btCapsuleShapeZ_new(radius, height))
 		{
 		}
