+++ BulletSharp/Collision/ConvexConcaveCollisionAlgorithm.cs
@@ -27,7 +27,7 @@ namespace BulletSharp
 			btConvexTriangleCallback_clearWrapperData(_native);
 		}
 
-		public void SetTimeStepAndCounters(float collisionMarginTriangle, DispatcherInfo dispatchInfo,
+		public void SetTimeStepAndCounters(double collisionMarginTriangle, DispatcherInfo dispatchInfo,
 			CollisionObjectWrapper convexBodyWrap, CollisionObjectWrapper triBodyWrap,
 			ManifoldResult resultOut)
 		{
@@ -85,7 +85,7 @@ namespace BulletSharp
 		[DllImport(Native.Dll, CallingConvention = Native.Conv), SuppressUnmanagedCodeSecurity]
 		static extern void btConvexTriangleCallback_setManifoldPtr(IntPtr obj, IntPtr value);
 		[DllImport(Native.Dll, CallingConvention = Native.Conv), SuppressUnmanagedCodeSecurity]
-		static extern void btConvexTriangleCallback_setTimeStepAndCounters(IntPtr obj, float collisionMarginTriangle, IntPtr dispatchInfo, IntPtr convexBodyWrap, IntPtr triBodyWrap, IntPtr resultOut);
+		static extern void btConvexTriangleCallback_setTimeStepAndCounters(IntPtr obj, double collisionMarginTriangle, IntPtr dispatchInfo, IntPtr convexBodyWrap, IntPtr triBodyWrap, IntPtr resultOut);
 		[DllImport(Native.Dll, CallingConvention = Native.Conv), SuppressUnmanagedCodeSecurity]
 		static extern void btConvexTriangleCallback_setTriangleCount(IntPtr obj, int value);
 	}
